backend/requirements.txt
# Heka Backend Requirements
# Python 3.11+

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Rate Limiting
slowapi==0.1.9

# Database
motor==3.3.2  # Async MongoDB driver
pymongo[srv]==4.6.1  # MongoDB Python driver with SRV support for Atlas
dnspython==2.4.2  # Required for mongodb+srv:// connections
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2  # Direct bcrypt instead of passlib
# passlib[bcrypt]==1.7.4  # Removed - using bcrypt directly
python-multipart==0.0.6

# AI Integration
openai>=1.40.0  # Updated for compatibility

# Environment & Config
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0  # Required for email validation

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0

# Payment Processing
stripe==7.2.0

